+++ packages/cli/src/commands/deploy/service.ts
@@ -1,13 +1,14 @@
 import { Command, flags } from '@oclif/command'
 import Listr from 'listr'
 import LCD from '@mesg/api/lib/lcd'
-import { compile, create, ICompile, ICreate } from '../../tasks/service'
 import Vault from '@mesg/vault'
 import FileStore from '@mesg/vault/lib/store/file'
 import { join } from 'path'
 import { parse } from 'url'
 import { prompt } from 'inquirer'
 import { Credential } from '../login'
+import { IService } from '@mesg/api/lib/service-lcd'
+import { compile, create } from '../../utils/service'
 
 const ipfsClient = require('ipfs-http-client')
 
@@ -25,30 +26,41 @@ export default class Service extends Command {
     default: './'
   }]
 
+  private ipfsClient = ipfsClient('ipfs.app.mesg.com', '5001', { protocol: 'http' })
+  private vault = new Vault<Credential>(new FileStore(join(this.config.configDir, 'credentials.json')))
+
   async run() {
     const { args, flags } = this.parse(Service)
-    const vault = new Vault<Credential>(new FileStore(join(this.config.configDir, 'credentials.json')))
+    const lcd = new LCD(flags.registry)
+
     const key = parse(flags.registry).hostname
-    if (!vault.contains(key)) this.error('no account found, please run `mesg-cli login`')
+    if (!this.vault.contains(key)) this.error('no account found, please run `mesg-cli login`')
     const password = flags.password
       ? flags.password
       : ((await prompt([{ name: 'password', type: 'password', message: 'Type the password of your account' }])) as any).password
 
-    const credential = vault.get(key, password)
+    const credential = this.vault.get(key, password)
 
-    const tasks = new Listr<ICompile | ICreate>([
-      compile,
-      create
+    let definition: IService
+    let service: IService
+
+    const tasks = new Listr([
+      {
+        title: 'Compiling service',
+        task: async () => {
+          definition = await compile(args.PATH, this.ipfsClient)
+        }
+      },
+      {
+        title: 'Creating service',
+        task: async () => {
+          service = await create(lcd, definition, credential.mnemonic)
+        }
+      },
     ])
-    const result = await tasks.run({
-      ipfsClient: ipfsClient('ipfs.app.mesg.com', '5001', { protocol: 'http' }),
-      path: args.PATH,
-      lcd: new LCD(flags.registry),
-      mnemonic: credential.mnemonic
-    })
+    await tasks.run()
 
-    const serviceHash = (result as ICreate).serviceHash
-    this.log(`Service deployed with the hash ${serviceHash}`)
-    this.log(`https://explorer.testnet.mesg.com/services/${serviceHash}`)
+    this.log(`Service deployed with the hash ${service.hash}`)
+    this.log(`https://explorer.testnet.mesg.com/services/${service.hash}`)
   }
 }